+++ CHANGELOG.md
@@ -18,6 +18,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   avoid going through the router every time to see if it needs to
   reconnect to said peers.
 
+### Removed 
+
+- Handshake when connecting to a new peer has been removed.
+
 ## [0.1.3] - 2023-11-22
 
 ### Added